+++ src/r3_core/src/bind.rs
@@ -172,9 +172,7 @@ use crate::{
     closure::Closure,
     hunk::Hunk,
     kernel::{self, cfg, prelude::*, raw, raw_cfg, StartupHook},
-    utils::{
-        refcell::RefCell, ComptimeVec, ConstAllocator, Init, PhantomInvariant, ZeroInit, Zeroable,
-    },
+    utils::{refcell::RefCell, ComptimeVec, ConstAllocator, Init, PhantomInvariant, Zeroable},
 };
 
 mod sorter;
@@ -470,8 +468,8 @@ impl<System>
     }
 
     /// Zero-initialize the binding contents.
-    pub const fn zeroed<T: ZeroInit>(self) -> BindDefiner<System, (), FnBindNever<T>> {
-        // Safety: `T: ZeroInit` means it's safe to zero-initialize
+    pub const fn zeroed<T: Zeroable>(self) -> BindDefiner<System, (), FnBindNever<T>> {
+        // Safety: `T: Zeroable` means it's safe to zero-initialize
         unsafe { self.zeroed_unchecked() }
     }